refactor: merge CommitMessageProcessor and MessageProcessor
